requirements-dev.txt
attrs==21.2.0
backports.entry-points-selectable==1.1.0
distlib==0.3.3
filelock==3.3.1
iniconfig==1.1.1
packaging==21.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
PyYAML==6.0
six==1.16.0
toml==0.10.2
tox==3.24.4
virtualenv==20.8.1
